Rename "dart-pkg" to "dart_pkg" in the Fuchsia Flutter Engine
This change renames engine/src/flutter/shell/platform/fuchsia/dart-pkg
to engine/src/flutter/shell/platform/fuchsia/dart_pkg. This is being
done since some build systems have trouble with directories
containing dashes ("-") instead of underscores ("_").
