+++ engine/src/flutter/shell/platform/fuchsia/BUILD.gn
@@ -90,7 +90,7 @@ if (enable_unittests) {
     testonly = true
 
     deps = [
-      "dart-pkg/zircon:tests",
+      "dart_pkg/zircon:tests",
       "dart_runner:tests",
       "flutter:tests",
     ]
