+++ engine/src/flutter/shell/platform/fuchsia/dart_pkg/fuchsia/sdk_ext/fuchsia.cc
@@ -2,18 +2,19 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/fuchsia/dart-pkg/fuchsia/sdk_ext/fuchsia.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/fuchsia/sdk_ext/fuchsia.h"
 
 #include <zircon/syscalls.h>
+#include <zircon/types.h>
 
 #include <cstdio>
 #include <cstring>
 #include <memory>
 #include <vector>
 
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon/sdk_ext/handle.h"
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon/sdk_ext/natives.h"
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon/sdk_ext/system.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/handle.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/natives.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/system.h"
 #include "third_party/dart/runtime/include/dart_api.h"
 #include "third_party/tonic/dart_binding_macros.h"
 #include "third_party/tonic/dart_class_library.h"
