+++ engine/src/flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/handle.h
@@ -11,7 +11,7 @@
 #include <vector>
 
 #include "flutter/fml/memory/ref_counted.h"
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon/sdk_ext/handle_waiter.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/handle_waiter.h"
 #include "third_party/dart/runtime/include/dart_api.h"
 #include "third_party/tonic/dart_library_natives.h"
 #include "third_party/tonic/dart_wrappable.h"
