+++ engine/src/flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/handle_disposition.h
@@ -10,7 +10,7 @@
 #include <vector>
 
 #include "flutter/fml/memory/ref_counted.h"
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon/sdk_ext/handle.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/handle.h"
 #include "third_party/dart/runtime/include/dart_api.h"
 #include "third_party/tonic/dart_library_natives.h"
 #include "third_party/tonic/dart_wrappable.h"
