+++ engine/src/flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/system.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon/sdk_ext/system.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/system.h"
 
 #include <array>
 
@@ -19,6 +19,8 @@
 #include <zircon/syscalls/iob.h>
 
 #include "flutter/fml/unique_fd.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/handle.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon/sdk_ext/handle_waiter.h"
 #include "third_party/tonic/dart_binding_macros.h"
 #include "third_party/tonic/dart_class_library.h"
