+++ engine/src/flutter/shell/platform/fuchsia/dart_pkg/zircon_ffi/clock.h
@@ -5,7 +5,7 @@
 #ifndef FLUTTER_SHELL_PLATFORM_FUCHSIA_DART_PKG_ZIRCON_FFI_CLOCK_H_
 #define FLUTTER_SHELL_PLATFORM_FUCHSIA_DART_PKG_ZIRCON_FFI_CLOCK_H_
 
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon_ffi/macros.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon_ffi/macros.h"
 
 #include <stdint.h>
