+++ engine/src/flutter/shell/platform/fuchsia/dart_pkg/zircon_ffi/dart_dl.cc
@@ -1,7 +1,7 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon_ffi/dart_dl.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon_ffi/dart_dl.h"
 
 #include "flutter/fml/logging.h"
 #include "include/dart_api_dl.h"
