+++ engine/src/flutter/shell/platform/fuchsia/dart_pkg/zircon_ffi/dart_dl.h
@@ -4,7 +4,7 @@
 #ifndef FLUTTER_SHELL_PLATFORM_FUCHSIA_DART_PKG_ZIRCON_FFI_DART_DL_H_
 #define FLUTTER_SHELL_PLATFORM_FUCHSIA_DART_PKG_ZIRCON_FFI_DART_DL_H_
 
-#include "flutter/shell/platform/fuchsia/dart-pkg/zircon_ffi/macros.h"
+#include "flutter/shell/platform/fuchsia/dart_pkg/zircon_ffi/macros.h"
 
 #include <stdint.h>
