+++ engine/src/flutter/shell/platform/fuchsia/dart_runner/kernel/libraries.yaml
@@ -19,12 +19,12 @@ dart_runner:
       uri: "../embedder/builtin.dart"
 
     zircon:
-      uri: "../../dart-pkg/zircon/lib/zircon.dart"
+      uri: "../../dart_pkg/zircon/lib/zircon.dart"
 
     zircon_ffi:
-      uri: "../../dart-pkg/zircon_ffi/lib/zircon_ffi.dart"
+      uri: "../../dart_pkg/zircon_ffi/lib/zircon_ffi.dart"
 
     fuchsia:
-      uri: "../../dart-pkg/fuchsia/lib/fuchsia.dart"
+      uri: "../../dart_pkg/fuchsia/lib/fuchsia.dart"
