+++ engine/src/flutter/shell/platform/fuchsia/flutter/BUILD.gn
@@ -176,8 +176,8 @@ template("runner_sources") {
              "${fuchsia_sdk}/pkg/trace-provider-so",
              "${fuchsia_sdk}/pkg/vfs_cpp",
              "${fuchsia_sdk}/pkg/zx",
-             "//flutter/shell/platform/fuchsia/dart-pkg/fuchsia",
-             "//flutter/shell/platform/fuchsia/dart-pkg/zircon",
+             "//flutter/shell/platform/fuchsia/dart_pkg/fuchsia",
+             "//flutter/shell/platform/fuchsia/dart_pkg/zircon",
            ] + flutter_deps
   }
 }
